moved test_batch_size to the mlir/test_features.py and broke it into two
