+++ forge/test/mlir/mnist/training/test_training.py
@@ -93,45 +93,3 @@ def test_mnist_training():
             break
 
     print(f"Test (total) loss: {test_loss}")
-
-@pytest.mark.parametrize("batch_size", [1, 2, 16])
-@pytest.mark.parametrize("in_features", [784])
-@pytest.mark.parametrize("out_features", [10])
-@pytest.mark.parametrize("enable_training", [True, False])
-def test_batch_size(batch_size, in_features, out_features, enable_training):
-    class SimpleModel(nn.Module):
-        def __init__(self):
-            super(SimpleModel, self).__init__()
-            self.linear = nn.Linear(in_features, out_features)
-
-        def forward(self, x):
-            y = self.linear(x)
-            return nn.functional.softmax(y, dim=-1)
-    
-    in_data = torch.rand(batch_size, in_features)
-    out_data = torch.randint(0, out_features, (batch_size,))
-    target = nn.functional.one_hot(out_data, num_classes=out_features).float()
-
-    model = SimpleModel()
-
-
-    loss_fn = torch.nn.CrossEntropyLoss() if enable_training else None
-    optimizer = torch.optim.SGD(model.parameters(), lr=0.001) if enable_training else None
-    tt_model = forge.compile(model, sample_inputs=[torch.rand(batch_size, in_features)], loss=loss_fn, optimizer=optimizer)
-
-    if enable_training:
-        optimizer.zero_grad()
-    
-    pred = tt_model(in_data)[0]
-    golden_pred = model(in_data)
-    assert compare_with_golden(golden_pred, pred, pcc=0.95)
-
-    if not enable_training:
-        return
-
-    loss = loss_fn(pred, target)
-    golden_loss = loss_fn(golden_pred, target)
-    assert torch.allclose(loss, golden_loss, rtol=1e-2)
-    
-    loss.backward()
-    tt_model.backward(pred.grad)
