+++ forge/test/mlir/test_features.py
@@ -10,7 +10,7 @@ import torch
 from torch import nn
 
 import forge
-from forge.op.eval.common import compare_with_golden_pcc
+from forge.op.eval.common import compare_with_golden_pcc, compare_with_golden
 
 
 def test_multiple_inputs():
@@ -90,3 +90,66 @@ def test_matmul_bias(batch_size, linear_features):
     co_out = [co.to("cpu") for co in co_out]
     fw_out = [fw_out] if isinstance(fw_out, torch.Tensor) else fw_out
     assert all([compare_with_golden_pcc(golden=fo, calculated=co, pcc=0.99) for fo, co in zip(fw_out, co_out)])
+
+
+@pytest.mark.parametrize("batch_size", [1, 2, 16])
+@pytest.mark.parametrize("in_features", [784])
+@pytest.mark.parametrize("out_features", [10])
+def test_batch_size_inference(batch_size, in_features, out_features):
+    class SimpleModel(nn.Module):
+        def __init__(self):
+            super(SimpleModel, self).__init__()
+            self.linear = nn.Linear(in_features, out_features)
+
+        def forward(self, x):
+            y = self.linear(x)
+            return nn.functional.softmax(y, dim=-1)
+    
+    in_data = torch.rand(batch_size, in_features)
+    out_data = torch.randint(0, out_features, (batch_size,))
+
+    model = SimpleModel()
+
+
+    tt_model = forge.compile(model, sample_inputs=[torch.rand(batch_size, in_features)])
+    
+    pred = tt_model(in_data)[0]
+    golden_pred = model(in_data)
+    assert compare_with_golden(golden_pred, pred, pcc=0.95)
+
+
+@pytest.mark.parametrize("batch_size", [1, 2, 16])
+@pytest.mark.parametrize("in_features", [784])
+@pytest.mark.parametrize("out_features", [10])
+def test_batch_size_training(batch_size, in_features, out_features):
+    class SimpleModel(nn.Module):
+        def __init__(self):
+            super(SimpleModel, self).__init__()
+            self.linear = nn.Linear(in_features, out_features)
+
+        def forward(self, x):
+            y = self.linear(x)
+            return nn.functional.softmax(y, dim=-1)
+    
+    in_data = torch.rand(batch_size, in_features)
+    out_data = torch.randint(0, out_features, (batch_size,))
+    target = nn.functional.one_hot(out_data, num_classes=out_features).float()
+
+    model = SimpleModel()
+
+    loss_fn = torch.nn.CrossEntropyLoss()
+    optimizer = torch.optim.SGD(model.parameters(), lr=0.001)
+    tt_model = forge.compile(model, sample_inputs=[torch.rand(batch_size, in_features)], loss=loss_fn, optimizer=optimizer)
+
+    optimizer.zero_grad()
+    
+    pred = tt_model(in_data)[0]
+    golden_pred = model(in_data)
+    assert compare_with_golden(golden_pred, pred, pcc=0.95)
+
+    loss = loss_fn(pred, target)
+    golden_loss = loss_fn(golden_pred, target)
+    assert torch.allclose(loss, golden_loss, rtol=1e-2)
+    
+    loss.backward()
+    tt_model.backward(pred.grad)
